Delete spelunk-cli's dead tests/common module
`crates/spelunk-cli/tests/common/mod.rs` had no importer and had not
compiled since the workspace split. It refers to a crate named `spelunk`
(`spelunk::storage::Database`, `spelunk::server::db::ServerDb`), but the
library is `spelunk_core` and `spelunk` is only the CLI binary name, so
adding `mod common;` to any test yields four E0433s.

Deleted rather than repaired. The crate already has a live shared test
module in `tests/plumbing_helpers.rs`; repairing this one would install a
competing convention, and its `open_test_server_db` would have pulled a
`spelunk-server` dependency in for zero callers. docs/testing.md already
stated that spelunk-cli keeps no independent copy of these fixtures.

Keeping it was the worse option for a second reason. It contained no
`isolate_git_config`, yet `scripts/check-git-isolation.sh` accepts a bare
`mod common;` as proof of git-config isolation, so any test wiring it in
would have satisfied the lint without gaining isolation. No spelunk-cli
test used it as that marker, so the lint is unaffected by the removal.

Its `register_sqlite_vec()` had meanwhile been copied inline into three
tests. That copy now lives once in `plumbing_helpers.rs` and the three
call sites share it, which is the deduplication a repair would have been
for, without the second fixture module.
